+++ modules/Student_Billing/functions.inc.php
@@ -405,6 +405,7 @@ function _makePaymentsFileInput( $value, $column )
  * Save Fees File
  *
  * @since 10.4
+ * @since 10.8.2 Add datetime to filename to make it harder to predict
  *
  * @param  int|string $id Fee ID or 'new'.
  *
@@ -422,12 +423,22 @@ function _saveFeesFile( $id )
 		return '';
 	}
 
+	$file_name_no_ext = no_accents( mb_substr(
+		$_FILES[ $input ]['name'],
+		0,
+		mb_strrpos( $_FILES[ $input ]['name'], '.' )
+	) );
+
+	$file_name_no_ext .= '_' . date( 'Y-m-d_His' );
+
 	$file_attached = FileUpload(
 		$input,
 		$FileUploadsPath . UserSyear() . '/student_' . UserStudentID() . '/',
 		FileExtensionWhiteList(),
 		0,
-		$error
+		$error,
+		'',
+		$file_name_no_ext
 	);
 
 	// Fix SQL error when quote in uploaded file name.