+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 10.8.2
 -----------------
 - PHP<7 Fix add microseconds to filename to make it harder to predict in StudentAssignments.fnc.php, thanks to @jeffreygaor
+- Add datetime to filename to make it harder to predict in Accounting/functions.inc.php & Student_Billing/functions.inc.php
 
 Changes in 10.8.1
 -----------------